pyproject.toml
[project]
name = "elab-doc-sync"
version = "0.1.0"
description = "Sync Markdown docs to eLabFTW items with image upload and diff-based updates"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28",
    "markdown>=3.4",
    "pyyaml>=6.0",
    "urllib3>=2.0",
]

[project.scripts]
elab-doc-sync = "elab_doc_sync.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/elab_doc_sync"]

[tool.hatch.build.targets.wheel.force-include]
"template" = "elab_doc_sync/template"
